Big logger and argument update

+++ src/system/time.cpp
@@ -1,16 +1,16 @@
-#include <LiteGL/time.hpp>
+#include <LiteGL/system/time.hpp>
 #include <ctime>
-#include <LiteGL/types.hpp>
+#include <LiteGL/system/types.hpp>
 
 namespace LiteAPI{
     Time getCurrentTime(){
         std::time_t now = std::time(nullptr);
         std::tm* localTime = std::localtime(&now);
-        return {localTime->tm_hour,localTime->tm_min,localTime->tm_sec};
+        return {(uint8)localTime->tm_hour,(uint8)localTime->tm_min,(uint8)localTime->tm_sec};
     }
     Date getCurrentDate(){
         std::time_t now = std::time(nullptr);
         std::tm* localTime = std::localtime(&now);
-        return {localTime->tm_year + 1900,localTime->tm_mon + 1,localTime->tm_mday};
+        return {(uint16)(localTime->tm_year) + (uint16)1900,(uint8)localTime->tm_mon + (uint8)1,(uint8)localTime->tm_mday};
     }
 }
